(PC-22640)[API] fix: use index to filter idAtProvider by venue
